add typed pattern
